finished tests for actions
